Task6 loaders and callback

+++ app/src/main/java/com/demo/mymovies/MainActivity.java
@@ -1,20 +1,25 @@
 package com.demo.mymovies;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModelProvider;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.DisplayMetrics;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.view.View;
+import android.widget.ProgressBar;
 import android.widget.Switch;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.demo.mymovies.adapters.MovieAdapter;
 import com.demo.mymovies.data.MainViewModel;
@@ -24,19 +29,30 @@ import com.demo.mymovies.utils.NetworkUtils;
 
 import org.json.JSONObject;
 
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity implements LoaderManager.LoaderCallbacks<JSONObject> {
 
     private RecyclerView recyclerViewPosters;
     private MovieAdapter movieAdapter;
     private Switch switchSort;
     private TextView textViewPopularity;
     private TextView textViewTopRated;
-    private Toast toastMessage;
+    private ProgressBar progressBarLoading;
 
     private MainViewModel viewModel;
+
+    private static final int LOADER_ID = 333;
+    private LoaderManager loaderManager;
+
+    private static int page = 1;
+    private static int methodOfSort;
+    private static boolean isLoading = false;
+
+    private static String lang;
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -63,25 +79,38 @@ public class MainActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
+    private int getColumnCount() {
+        DisplayMetrics displayMetrics = new DisplayMetrics();
+        // теперь получаем хар-ки экрана и помещаем в объект DisplayMetrics
+        getWindowManager().getDefaultDisplay().getMetrics(displayMetrics);
+        // получаем аппаратно независимые пиксели, разделив на плотность (displayMetrics.density)
+        int width = (int) (displayMetrics.widthPixels / displayMetrics.density);
+        return Math.max(width / 185, 2);
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        lang = Locale.getDefault().getLanguage();
+        loaderManager = LoaderManager.getInstance(this);
         viewModel = new ViewModelProvider(this).get(MainViewModel.class);
         switchSort = findViewById(R.id.switchSort);
         textViewPopularity = findViewById(R.id.textViewPopularity);
         textViewTopRated = findViewById(R.id.textViewTopRated);
+        progressBarLoading = findViewById(R.id.progressBarLoading);
         recyclerViewPosters = findViewById(R.id.recyclerViewPosters);
-        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, 2));
+        recyclerViewPosters.setLayoutManager(new GridLayoutManager(this, getColumnCount()));
         movieAdapter = new MovieAdapter();
         recyclerViewPosters.setAdapter(movieAdapter);
 
         switchSort.setChecked(true);
         switchSort.setOnCheckedChangeListener((compoundButton, isChecked) -> {
+            page = 1;
             setMethodOfSort(isChecked);
         });
         switchSort.setChecked(false);
-        downloadData(NetworkUtils.POPULARITY, 1);
+        downloadData(NetworkUtils.POPULARITY, 1, lang);
 
         movieAdapter.setOnPosterClickListener((position -> {
             Movie movie = movieAdapter.getMovies().get(position);
@@ -91,11 +120,14 @@ public class MainActivity extends AppCompatActivity {
         }));
 
         movieAdapter.setOnReachEndListener(() -> {
-            if (toastMessage != null) {
-                toastMessage.cancel();
+            if (!isLoading) {
+//                if (toastMessage != null) {
+//                    toastMessage.cancel();
+//                }
+//                toastMessage = Toast.makeText(this, "Конец списка", Toast.LENGTH_SHORT);
+//                toastMessage.show();
+                downloadData(methodOfSort, page, lang);
             }
-            toastMessage = Toast.makeText(this, "Конец списка", Toast.LENGTH_SHORT);
-            toastMessage.show();
         });
 
         textViewPopularity.setOnClickListener((view) -> {
@@ -110,11 +142,15 @@ public class MainActivity extends AppCompatActivity {
 
         LiveData<List<Movie>> moviesFromLiveData = viewModel.getMovies();
         // Теперь каждый раз, когда данные в БД будут меняться, мы их будем устанавливать у адаптера
-        moviesFromLiveData.observe(this, movies -> movieAdapter.setMovies(movies));
+        moviesFromLiveData.observe(this, movies -> {
+                    if (page == 1) {
+                        movieAdapter.setMovies(movies);
+                    }
+                }
+        );
     }
 
     private void setMethodOfSort(boolean isTopRated) {
-        int methodOfSort;
         if (isTopRated) {
             textViewTopRated.setTextColor(getResources().getColor(R.color.colorAccent));
             textViewPopularity.setTextColor(getResources().getColor(R.color.white));
@@ -124,20 +160,51 @@ public class MainActivity extends AppCompatActivity {
             textViewPopularity.setTextColor(getResources().getColor(R.color.colorAccent));
             textViewTopRated.setTextColor(getResources().getColor(R.color.white));
         }
-        downloadData(methodOfSort, 1);
+        downloadData(methodOfSort, page, lang);
     }
 
-    private void downloadData(int methodOfSort, int page) {
-        JSONObject jsonObject = NetworkUtils.getJSONFromNetwork(methodOfSort, 1);
+    private void downloadData(int methodOfSort, int page, String lang) {
+        URL url = NetworkUtils.buildURL(methodOfSort, page, lang);
+        Bundle bundle = new Bundle();
+        bundle.putString("url", url.toString());
+        loaderManager.restartLoader(LOADER_ID, bundle, this);
+    }
+
+    @NonNull
+    @Override
+    public Loader<JSONObject> onCreateLoader(int id, @Nullable Bundle bundle) {
+        NetworkUtils.JSONLoader jsonLoader = new NetworkUtils.JSONLoader(this, bundle);
+        jsonLoader.setOnStartLodingListener(() -> {
+            progressBarLoading.setVisibility(View.VISIBLE);
+            isLoading = true;
+        });
+        return jsonLoader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<JSONObject> loader, JSONObject jsonObject) {
         ArrayList<Movie> movies = JSONUtils.getMoviesFromJSON(jsonObject);
         if (movies != null && !movies.isEmpty()) {
-            viewModel.deleteAllMovies();
+            if (page == 1) {
+                viewModel.deleteAllMovies();
+                movieAdapter.clear();
+            }
             try {
                 Thread.sleep(10);
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
             viewModel.insertListMovies(movies);
+            movieAdapter.addMovies(movies);
+            page++;
         }
+        isLoading = false;
+        progressBarLoading.setVisibility(View.INVISIBLE);
+        loaderManager.destroyLoader(LOADER_ID);
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<JSONObject> loader) {
+
     }
 }
